+++ test/test_store.py
@@ -34,7 +34,7 @@ To create a new unit test of this kind,
    of the subdirectory is arbitrary, but it is useful to give it the same
    name as the unit test method.
 4. Remove the _save_request_to and _save_file_to arguments from the open_data
-   call, and add a 'client=CDSClientMock' argument to the CDSDataOpener
+   call, and add a 'client_class=CDSClientMock' argument to the CDSDataOpener
    constructor.
 """
 
@@ -55,11 +55,16 @@ from xcube_cds.store import CDSDataOpener
 from xcube_cds.store import CDSDataStore
 from xcube_cds.store import CDSDatasetHandler
 
+_CDS_API_URL = 'dummy'
+_CDS_API_KEY = 'dummy'
+
 
 class CDSStoreTest(unittest.TestCase):
 
     def test_open(self):
-        opener = CDSDataOpener(client=CDSClientMock)
+        opener = CDSDataOpener(client_class=CDSClientMock,
+                               cds_api_url=_CDS_API_URL,
+                               cds_api_key=_CDS_API_KEY)
         dataset = opener.open_data(
             'reanalysis-era5-single-levels-monthly-means:'
             'monthly_averaged_reanalysis',
@@ -76,7 +81,9 @@ class CDSStoreTest(unittest.TestCase):
         self.assertEqual(10, len(dataset.variables['time']))
 
     def test_normalize_variable_names(self):
-        store = CDSDataStore(client=CDSClientMock, normalize_names=True)
+        store = CDSDataStore(client_class=CDSClientMock, normalize_names=True,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = store.open_data(
             'reanalysis-era5-single-levels-monthly-means:'
             'monthly_averaged_reanalysis',
@@ -90,7 +97,8 @@ class CDSStoreTest(unittest.TestCase):
         self.assertTrue('p54_162' in dataset.variables)
 
     def test_invalid_data_id(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         with self.assertRaises(ValueError):
             store.open_data(
                 'this-data-id-does-not-exist',
@@ -99,7 +107,8 @@ class CDSStoreTest(unittest.TestCase):
             )
 
     def test_request_parameter_out_of_range(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         with self.assertRaises(ValidationError):
             store.open_data(
                 'reanalysis-era5-single-levels:ensemble_mean',
@@ -111,7 +120,9 @@ class CDSStoreTest(unittest.TestCase):
             )
 
     def test_era5_land_monthly(self):
-        store = CDSDataStore(client=CDSClientMock)
+        store = CDSDataStore(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = store.open_data(
             'reanalysis-era5-land-monthly-means:'
             'monthly_averaged_reanalysis',
@@ -126,7 +137,9 @@ class CDSStoreTest(unittest.TestCase):
         self.assertTrue('u10' in dataset.variables)
 
     def test_era5_single_levels_hourly(self):
-        store = CDSDataStore(client=CDSClientMock)
+        store = CDSDataStore(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = store.open_data(
             'reanalysis-era5-single-levels:'
             'reanalysis',
@@ -142,7 +155,9 @@ class CDSStoreTest(unittest.TestCase):
         self.assertEqual(26, len(dataset.variables['time']))
 
     def test_era5_land_hourly(self):
-        store = CDSDataStore(client=CDSClientMock)
+        store = CDSDataStore(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = store.open_data(
             'reanalysis-era5-land',
             variable_names=['2m_temperature'],
@@ -157,7 +172,9 @@ class CDSStoreTest(unittest.TestCase):
         self.assertEqual(26, len(dataset.variables['time']))
 
     def test_era5_bounds(self):
-        opener = CDSDataOpener(client=CDSClientMock)
+        opener = CDSDataOpener(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = opener.open_data(
             'reanalysis-era5-single-levels-monthly-means:'
             'monthly_averaged_reanalysis',
@@ -179,7 +196,9 @@ class CDSStoreTest(unittest.TestCase):
         self.assertLessEqual(south, north)
 
     def test_soil_moisture(self):
-        store = CDSDataStore(client=CDSClientMock)
+        store = CDSDataStore(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         data_id = 'satellite-soil-moisture:volumetric:monthly'
         dataset = store.open_data(
             data_id,
@@ -200,7 +219,8 @@ class CDSStoreTest(unittest.TestCase):
                          sorted(map(str, dataset.data_vars)))
 
     def test_list_and_describe_data_ids(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         data_ids = store.get_data_ids()
         self.assertIsInstance(data_ids, Iterator)
         for data_id in data_ids:
@@ -212,7 +232,8 @@ class CDSStoreTest(unittest.TestCase):
             self.assertIsInstance(descriptor, DataDescriptor)
 
     def test_open_data_empty_variables_list_1(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = store.open_data(
             'reanalysis-era5-land-monthly-means:monthly_averaged_reanalysis',
             variable_names=[],
@@ -226,7 +247,8 @@ class CDSStoreTest(unittest.TestCase):
         self.assertEqual(361, len(dataset.variables['lon']))
 
     def test_open_data_empty_variables_list_2(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         dataset = store.open_data(
             'satellite-soil-moisture:volumetric:10-day',
             variable_names=[],
@@ -241,7 +263,9 @@ class CDSStoreTest(unittest.TestCase):
         self.assertEqual(1441, len(dataset.variables['lon']))
 
     def test_open_data_null_variables_list(self):
-        store = CDSDataStore(client=CDSClientMock)
+        store = CDSDataStore(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         data_id = 'reanalysis-era5-single-levels-monthly-means:'\
             'monthly_averaged_reanalysis'
         schema = store.get_open_data_params_schema(data_id)
@@ -257,7 +281,8 @@ class CDSStoreTest(unittest.TestCase):
         self.assertEqual(n_vars, len(dataset.data_vars))
 
     def test_era5_describe_data(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         descriptor = store.describe_data(
             'reanalysis-era5-single-levels:reanalysis')
         self.assertEqual(265, len(descriptor.data_vars))
@@ -283,7 +308,8 @@ class CDSStoreTest(unittest.TestCase):
             CDSDatasetHandler.convert_time_range([])  # incorrect list length
 
     def test_get_open_params_schema_without_data_id(self):
-        opener = CDSDataOpener()
+        opener = CDSDataOpener(cds_api_url=_CDS_API_URL,
+                               cds_api_key=_CDS_API_KEY)
         schema = opener.get_open_data_params_schema()
 
         actual = schema.to_dict()
@@ -299,19 +325,23 @@ class CDSStoreTest(unittest.TestCase):
         )
 
     def test_search_data_invalid_id(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         with self.assertRaises(DataStoreError):
             store.search_data('This is an invalid ID.')
 
     def test_search_data_valid_id(self):
-        store = CDSDataStore()
+        store = CDSDataStore(cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         # The CDS API doesn't offer a search function, so "not implemented"
         # is expected here.
         with self.assertRaises(NotImplementedError):
             store.search_data('dataset')
 
     def test_copy_on_open(self):
-        store = CDSDataStore(client=CDSClientMock)
+        store = CDSDataStore(client_class=CDSClientMock,
+                             cds_api_url=_CDS_API_URL,
+                             cds_api_key=_CDS_API_KEY)
         data_id = 'satellite-soil-moisture:volumetric:monthly'
         with tempfile.TemporaryDirectory() as temp_dir:
             request_path = os.path.join(temp_dir, 'request.json')
@@ -370,6 +400,136 @@ class CDSStoreTest(unittest.TestCase):
             CDSDataStore().get_open_data_params_schema(),
             xcube.util.jsonschema.JsonObjectSchema
         )
+
+
+class ClientUrlTest(unittest.TestCase):
+
+    """Tests connected with passing CDS API URL and key to opener or store."""
+
+    def setUp(self):
+        self.old_environment = dict(os.environ)
+        self.temp_dir = tempfile.TemporaryDirectory()
+
+    def tearDown(self):
+        os.environ.clear()
+        os.environ.update(self.old_environment)
+        self.temp_dir.cleanup()
+
+    def test_client_url_and_key_parameters(self):
+        """Test passing URL and key parameters to client constructor
+
+        This test verifies that the CDS API URL and key, when specified as
+        parameters to CDSDataOpener, are correctly passed to the CDS client,
+        overriding any configuration file or environment variable settings.
+        """
+
+        self._set_up_api_configuration('wrong URL 1', 'wrong key 1',
+                                       'wrong URL 2', 'wrong key 2')
+        cds_api_url = 'https://example.com/'
+        cds_api_key = 'xyzzy'
+        opener = CDSDataOpener(client_class=CDSClientMock,
+                               cds_api_url=cds_api_url,
+                               cds_api_key=cds_api_key)
+        opener.open_data(
+            'reanalysis-era5-single-levels-monthly-means:'
+            'monthly_averaged_reanalysis',
+            variable_names=['2m_temperature'],
+            bbox=[-180, -90, 180, 90],
+            spatial_res=0.25,
+            time_period='1M',
+            time_range=['2015-10-15', '2015-10-15'],
+        )
+        client = opener.last_instantiated_client
+        self.assertEqual(cds_api_url, client.url)
+        self.assertEqual(cds_api_key, client.key)
+
+    def test_client_url_and_key_environment_variables(self):
+        """Test passing URL and key parameters via environment variables
+
+        This test verifies that the CDS API URL and key, when specified in
+        environment variables, are correctly passed to the CDS client,
+        overriding any configuration file settings.
+        """
+
+        cds_api_url = 'https://example.com/'
+        cds_api_key = 'xyzzy'
+        self._set_up_api_configuration('wrong URL 1', 'wrong key 1',
+                                       cds_api_url, cds_api_key)
+        client = self._get_client()
+        self.assertEqual(cds_api_url, client.url)
+        self.assertEqual(cds_api_key, client.key)
+
+    def test_client_url_and_key_rc_file(self):
+        """Test passing URL and key parameters via environment variables
+
+        This test verifies that the CDS API URL and key, when specified in
+        a configuration file, are correctly passed to the CDS client.
+        """
+
+        cds_api_url = 'https://example.com/'
+        cds_api_key = 'xyzzy'
+        self._set_up_api_configuration(cds_api_url, cds_api_key)
+        opener = CDSDataOpener(client_class=CDSClientMock)
+        opener.open_data(
+            'reanalysis-era5-single-levels-monthly-means:'
+            'monthly_averaged_reanalysis',
+            variable_names=['2m_temperature'],
+            bbox=[-180, -90, 180, 90],
+            spatial_res=0.25,
+            time_period='1M',
+            time_range=['2015-10-15', '2015-10-15'],
+        )
+        client = opener.last_instantiated_client
+        self.assertEqual(cds_api_url, client.url)
+        self.assertEqual(cds_api_key, client.key)
+
+    @staticmethod
+    def _get_client(**opener_args):
+        """Return the client instantiated to open a dataset
+
+        Open a dataset and return the client that was instantiated to execute
+        the CDS API query.
+        """
+        opener = CDSDataOpener(client_class=CDSClientMock, **opener_args)
+        opener.open_data(
+            'reanalysis-era5-single-levels-monthly-means:'
+            'monthly_averaged_reanalysis',
+            variable_names=['2m_temperature'],
+            bbox=[-180, -90, 180, 90],
+            spatial_res=0.25,
+            time_period='1M',
+            time_range=['2015-10-15', '2015-10-15'],
+        )
+        return opener.last_instantiated_client
+
+    def _set_up_api_configuration(self, url_rc, key_rc,
+                                  url_env=None, key_env=None):
+        """Set up a configuration file and, optionally, environment variables
+
+        The teardown function will take care of the clean-up.
+
+        :param url_rc: API URL to be written to configuration file
+        :param key_rc: API key to be written to configuration file
+        :param url_env: API URL to be written to environment variable
+        :param key_env: API key to be written to environment variable
+        :return: an instantiated CDS client object
+        """
+        rc_path = os.path.join(self.temp_dir.name, "cdsapi.rc")
+        with open(rc_path, 'w') as fh:
+            fh.write(f'url: {url_rc}\n')
+            fh.write(f'key: {key_rc}\n')
+        ClientUrlTest._erase_environment_variables()
+        os.environ['CDSAPI_RC'] = rc_path
+        if url_env is not None:
+            os.environ['CDSAPI_URL'] = url_env
+        if key_env is not None:
+            os.environ['CDSAPI_KEY'] = key_env
+
+    @staticmethod
+    def _erase_environment_variables():
+        for var in 'CDSAPI_URL', 'CDSAPI_KEY', 'CDSAPI_RC':
+            if var in os.environ:
+                del os.environ[var]
 
 
 if __name__ == '__main__':